fix: Darken the image in About page
In ~/themes/airspace-hugo/layouts/about, list.html file creats the About page.
I wanted to make the image in the About picture not to be faded, which was the default setting.
In list.html, I found out that the image is defined as an "about-img" class.
I changed it to be a "block" class, and the image did not fade.